requirements.txt
accelerate==0.33.0
transformers==4.44.0
sentencepiece==0.2.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
requests==2.32.3
tqdm==4.66.5